+++ examples/reddit-clone/src/workflows.rs
@@ -9,8 +9,8 @@ use std::time::Duration;
 
 use autumn_harvest::error::database_error;
 use autumn_harvest::prelude::*;
+use autumn_harvest_plugin::{AppDbPool, WorkflowStartRequest};
 use autumn_web::AppState;
-use autumn_web_harvest::{AppDbPool, WorkflowStartRequest};
 use diesel::ExpressionMethods;
 use diesel::QueryDsl;
 use diesel_async::RunQueryDsl;